# Improve performance for reading, writing, and iterating sequences.