Uses binary search to find data nodes corresponding to an index value and insertion position for new nodes to be added to the sequence.

As an example, I had a Sequence of 17000 transforms on my computer. Reading from *.seq.mha file takes 16 seconds now, compared to 85 seconds previously. Iterating through and computing performance metrics on it takes 15 seconds now, compared to 38 seconds previously.